refactor: CheckOrderProceeding 이벤트 구현
